Refactor: move properties and data sources to qae folder, update imports, and remove old files
