+++ src/qae/index.js
@@ -1,0 +1,4 @@
+import properties from './object-properties';
+import data from './data';
+
+export { properties, data };
