+++ src/qae/object-properties.js
@@ -1,4 +1,4 @@
-import pkg from '../package.json';
+import pkg from '../../package.json';
 
 const properties = {
   title: `${pkg.name} v${pkg.version}`,
